dt_model/requirements.txt
sklearn
cython
gcloud
google-cloud-datastore
schedule
google-cloud-storage